Create search results messaage as a separate component
